[feat, 407] filter experiment results from db
This is an initial implementation for filtering results by algorithm, scenario, dataset, syscall direction and features.
For features it traverses the dependency graph stored in the db and checks weather a path containing the specified feature per algorithm exits. This is done by the use of custom js functions in mongodb